+++ src/utils/aiAccountant/index.js
@@ -25,6 +25,7 @@ export { runOrchestrator, runAuditAgent, auditFromIssues, POLICY_VERSION } from 
 export { analyzePostedEntries, primaryAccount, isFlagged } from './analyst.js';
 export { runBooksAudit } from './booksAudit.js';
 export { buildCloseChecklist, buildComplianceCalendar } from './closeChecklist.js';
+export { proposeDepreciation, DEP_RULES } from './depreciation.js';
 export {
   resolveAccount,
   buildRunningLedger,